pyproject.toml
[project]
name = "sigil"
version = "1.0.0"
description = "Sigil Solver"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "dishka>=1.6.0",
    "fastapi-pagination>=0.13.3",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
    "numpy==1.26.4",
    "onnx>=1.14.0",
    # GPU build only for Linux/Windows + Python <3.12
    'onnxruntime-gpu; (platform_system == "Linux" or platform_system == "Windows") and python_version < "3.12"',
    'onnxruntime-silicon; platform_system == "Darwin" and platform_machine == "arm64"',
    # ONNX Runtime (choose per platform)
    'onnxruntime; (platform_system == "Darwin" and platform_machine == "x86_64") or (platform_system != "Darwin")',
    "pillow-heif>=1.1.0",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "scalar-fastapi>=1.2.3",
    "shortuuid>=1.0.13",
    "starlette-context>=0.4.0",
    "torch==2.2.2",
    "typer>=0.16.1",
    "ultralytics==8.3.111",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "autoflake>=2.3.1,<3",
    "black>=25.1.0,<26",
    "deptry>=0.23.0,<0.24",
    "flake8>=7.2.0,<8",
    "icecream>=2.1.5",
    "isort>=6.0.1,<7",
    "mypy>=1.15.0,<2",
    "pre-commit==4.2.0",
    "pyinstaller>=6.15.0",
    "pytest-asyncio>=0.23.7,<0.24",
    "pytest-cov>=5.0.0,<6",
    "pytest>=8.2.3,<9",
    "ruff>=0.11.8,<0.12",
    "uv-sort>=0.6.1",
]

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py312']
